ashmem: int/size_t and slight modernization.
ashmem_get_size_region() should return size_t, since
ashmem_create_region() takes a size_t.

This is a no-op (other than sign) for LP32, since size_t is 32 bits.

For LP64, this is an ABI change, but in effect it just means that we
don't zero the top 32 bits of the result register; any code that worked
before still works, and (highly unlikely) code that allocated > 2GiB
ashmem regions can now happily query that allocated size *provided*
that the implementation is memfd, not ashmem. The ashmem implementation
can only return an int from the kernel anyway, so it's already
truncated before it gets to this point :-(

Also lose bogus "\n"s from logging, and use unique_fd rather than goto.

Bug: http://b/205348471
Test: treehugger

Former-commit-id: 4f1a0c84ff5d897fc7c44ed1ec8e34faf6fd0b62

+++ libcutils/ashmem-host.cpp
@@ -17,8 +17,13 @@
 #include <cutils/ashmem.h>
 
 /*
- * Implementation of the user-space ashmem API for the simulator, which lacks
- * an ashmem-enabled kernel. See ashmem-dev.c for the real ashmem-based version.
+ * Implementation of the userspace ashmem API for the host, which lacks
+ * an ashmem-enabled kernel.
+ *
+ * See ashmem-dev.cpp for the real ashmem-based version.
+ *
+ * TODO: is this actually used on Darwin? If not, we can use the memfd_create()
+ * based implementation that's in ashmem-dev now...
  */
 
 #include <errno.h>
@@ -85,12 +90,10 @@ int ashmem_unpin_region(int /*fd*/, size_t /*offset*/, size_t /*len*/) {
     return 0 /*ASHMEM_IS_UNPINNED*/;
 }
 
-int ashmem_get_size_region(int fd)
-{
+size_t ashmem_get_size_region(int fd) {
     struct stat buf;
     if (!ashmem_validate_stat(fd, &buf)) {
         return -1;
     }
-
     return buf.st_size;
 }
